Rename error event to liveSyncError
Rename the error event to liveSyncError as EventEmitter expects instance of Error when event named `error` is raised.
So it fails to parse our object. Also it automatically raised Uncaught error in CLI when noone has added a handler for error event:`Uncaught, unspecified "error" event. ([object Object])`.
Fix incorrect message in an error where deviceIdentifier was not printed correctly.